requirements.txt
asgiref==3.5.2
Django==4.1.3
djangorestframework==3.14.0
et-xmlfile==1.1.0
numpy==1.23.4
openpyxl==3.0.10
pandas==1.5.1
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
xlrd==2.0.1
XlsxWriter==3.0.3
